azure-providers.js: Add AzureCarts for managing order creation
Users can have several in-progress carts, and we use the cart
information a lot.  This commit adds a factory that tracks a person's
carts using a Carts object.  The factory has an internal cart_sets
variable to keep references to any loaded carts, so multiple calls to:

  AzureCarts(person_id)

will all return the same Carts instance.  That makes it easy to share
state between several controllers and directives, and lets us avoid
the additional network requests that you'd get if each controller
loaded it's own cart data directly.  It also makes it easy to share
information about the currently-selected cart.

The goal is to have the next-to-cutoff cart be the default cart (this
commit just uses .carts[0]).  We store the currently-selected cart in
Carts.cart, and you can change the selection with
Carts.selectCart(order_id), since Benjamin doesn't want it to persist
between refreshes.  I'd consider having it persist between refreshes
and be bound to the current session, but then we'd need more API
endpoints to set that for the session.