azure-providers.js: Add Category.path(category)
Sluggified path using category names to show where a category stands
in the hierarchy.  For example:

  Healthy Foods → Bulk → Chocolate & Carob → Cocoa, Cacao, & Chocolate → Chocolate

will have a path of:

  healthy-foods/bulk/chocolate-carob/cocoa-cacao-chocolate/chocolate

We considered removing duplicate words (e.g. 'Chocolate' from 'Cocoa,
Cacao, & Chocolate' is already mentioned in 'Chocolate & Carob'), but
that is hard to do right without knowing your siblings.  For example,
we don't want to collapse both:

  Healthy Foods → Bulk → Chocolate & Carob → Baking Chips → Carob Chips
  Healthy Foods → Bulk → Chocolate & Carob → Baking Chips → Chocolate Chips

to:

  healthy-foods/bulk/chocolate-carob/baking-chips/chips

We could (for each slug in the path) query for siblings and collapse
accordingly, but I don't think auto-collapsing slugs is worth the
extra network access that would require.