azure-providers.js: Add AzureAPI.<model>.count
To figure out how many items match a given query.  I couldn't figure
out where to attach the data as part of an AzureAPI.<model>.query
request, so I added this new method.  Not as efficient as extracting
the count from the query, but it will get the job done.

I use the response interceptor [1,2] to extract the 'Count' header
from the response and attach it to the $request instance (we're using
HEAD, so there's no data in the response body).  We need to keep the
$request object, because the mutable object is already attached to the
$scope (or held somehow by the caller), so I attach the value under a
'count' attribute ($scope.favorites_count.count in example.html).

[1]: https://docs.angularjs.org/api/ngResource/service/$resource#usage
[2]: https://docs.angularjs.org/api/ng/service/$http#interceptors