azure-providers.js: create, save, and delete actions for $resources
Describe them in the README and exercise them in example.html.

_payload_headers holds headers used for requests with payloads (POSTs
and PUTs).  We take the union of the generic _headers and the
payload-specific _payload_headers for payload actions.

I've also added an @<identifier> (e.g. @id or @slug) to the
$resource's default parameters so it pulls that information from the
model data [1].

We can't use line.$save() directly, because there's no way to
automatically pull the price out of the model (sometimes we *do* want
to force a price), and usual users saving with prices will get:

  {"message": "cannot set price while the order is still in the cart",
   "code": 400}

You probably don't actually want to create or delete order-lines
directly, since that will invalidate the lines cached inside
AzureCarts.  We'll want a follow-up commit that updates Cart to add
line creation and removal methods to Cart that *do* update the cached
line lists appropriately.

[1]: https://docs.angularjs.org/api/ngResource/service/$resource#usage