azure-providers.js: Category.children hit cache before checking .category
We might have those children in the cache even if this.category hasn't
been setup yet.  For example, if we just instantiated a new Category
via AzureCategory for a category who's children we already fetched.
This happens in the category manager, where we work down through the
category tree (fetching children) to get to a particular category.  If
we delete the category, we also want to remove it from the parent's
child list, and it's convenient to use:

  var parent = node.ancestors[1];
  var p = new AzureCategory(parent);
  var children = p.children();
	var index = children.indexOf(node);
	children.splice(index, 1);

That doesn't give time for p.$promise.category to resolve, so the old
code would just return [] for p.children().  The new code
optimistically checks the cache, finds the children in the cache, and
returns them without needing to wait for p.$promise.category.