azure-providers.js: Add OrderLine.save()
I've also added OrderLine.cart, which we need a reference to so we can
update the totals (via the newly-factored-out Cart._calculateTotals)
after getting the new object.  We also recalculate the price for the
line, via the newly-factored-out OrderLine._calculatePrice).

The:

  ng-model-options="{updateOn: 'blur'}"

bit avoids excessive saves by only updating the model when the input
loses focus [1]:

  updateOn: string specifying which event should the input be bound
    to. You can set several events using an space delimited
    list. There is a special event called default that matches the
    default events belonging of the control.

and then later there's the debounce example [1]:

  ng-model-options="{ updateOn: 'default blur', debounce: {'default': 500, 'blur': 0} }"

The ng-change docs have [2]:

   The ngChange expression is only evaluated when a change in the
   input value causes a new value to be committed to the model.

[1]: https://docs.angularjs.org/api/ng/directive/ngModelOptions#usage
[2]: https://docs.angularjs.org/api/ng/directive/ngChange