azure-providers.js: Add AzureAPI.trip.mail()
This should be a temporary stop-gap (so there are no docs), since once
the new website is live it will be easy for customers to check the
expected delivery time for any stops they're associated with.
However, until that happens, the logistics folks need a way to mail
scheduling information to customers, and this will let them do that
independently.