azure-providers.js: Add .$promise to Category.children() response
This gives folks a hook to manipulate the children if they need to do
any post-processing after they've been populated.  We need this for
the internal website's category manager, to give the category
instances the attributes needed by Angular UI Tree.

I'm using $q.all [1] to get a promise that returns after the
child-listing request completes *and* after the individual
Category-setup requests spawned by the child-listing requests have
completed.  We need the initial child-listing request in the array of
promises to ensure the $q.all promise doesn't complete before we've
pushed any of the children's Category-setup promises.  All of the
promises return the same _children array, so we just grab the result
from the first promise and return that when children().$promise
resolves.

[1]: https://docs.angularjs.org/api/ng/service/$q#all