AzureAPI: Pass $http.post data through directly
Instead of using function arguments arguments for the first level of
keys.  This way you don't have to remember/lookup the argument order.

This is a backwards-incompatible change to those APIs, but it allows
for more direct binding between the API spec and the caller.  On the
one hand, that means we don't need to update this library if the API
shuffles its expected fields.  On the other hand, that means we can't
buffer callers when the API shuffles its expected fields (but
shuffling will probably include semantic changes, and callers will
have to update to handle those anyway).

The wrapper function taking postConfig and config gives us a new scope
to lock in their values, because JavaScript's blocks don't create a
new scope [1].  Otherwise all of the methods created from _posts will
have the URL and config of the final _posts entry (currently
resendConfirmationEmail).

[1]: http://stackoverflow.com/a/643664