AzureAPI: Add Algolia client indexes to models
This adds default indexes to four models that we've already pushed to
Algolia.  setAlgoliaIndexNames makes it easy for developers to
redirect those requests to alternative indexes while they test a new
model structure, or for them to additional indexes when they're
developing something new.

algoliasearch 3.8.1 is the current release (2015-09-21 [1]), but a
reference to algolia.Client landed in 2.9.0 [2], a reference to
initIndex landed in 2.0.0 [3], so anything in the 3.x series should
work for us (although obviously the more recent releases are likely to
be better maintained).

[1]: https://github.com/algolia/algoliasearch-client-js/releases
[2]: https://github.com/algolia/algoliasearch-client-js/commit/0dfff897bb9b8b3c8f8d43e3a6ee202b444a944c#diff-230865f0ac77bd3ba60bfbdf88011c9aR42
[3]: https://github.com/algolia/algoliasearch-client-js/commit/ae7727398d7563ea5cc46864d2291402a1d5c2b9#diff-04c6e90faac2675aa89e2176d2eec7d8R52