AzureAPIProvider: Change 'url' to 'setUrl'
Using the same name raised concerns about whether or not calling the
setter would clobber the setter itself, besides updating the
provider-wide variable.  I tested it with:

  var _this = this;
  this.url = function(value) {
    url = value;
    console.log('url', _this.url);
  };

Which logged:

  12:04:57.433 "url" function AzureAPIProvider/this.url()  1 azure-providers.js:54:12

so it looks like it it wasn't clobbering this.url.  But it's still
good to avoid that potential confusion, so this commit choses a
different name for the setter.