AzureProduct: Skip getObjectPromise for package-code lookups
When instantiating AzureProduct with a packaged product code, algolia
is queried by that code, and if no results are returned, then the API
is queried.  If either query returns exactly one result, that result
is the product needed.  However, instead of just returning this
result, only the product id is used to query algolia again via
AzureProduct's internal ObjectPromiseCache instance.

This fix eliminates the unnecessary second query and just returns the
product from the first query result.